requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
model-bakery==1.19.5
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
